dfsd

+++ organizations/models/organization_landing_page.py
@@ -16,7 +16,7 @@ class OrganizationLandingPage(models.Model):
     education_language = models.ForeignKey(EducationLanguage, on_delete=models.CASCADE, null=True)
     deleted = models.BooleanField(default=False)
     grant = models.BooleanField(default=False)
-    price = models.DecimalField(max_digits=10, decimal_places=2, null=True)
+    price = models.BigIntegerField(null=True)
     requirements = models.TextField(null=True, blank=True)
     shift = models.ForeignKey(Shift, on_delete=models.SET_NULL, null=True, blank=True)
     field = models.ForeignKey("organizations.OrganizationFields", on_delete=models.SET_NULL, null=True, blank=True)
